+++ aoc2020/src/lib.rs
@@ -8,7 +8,7 @@ pub fn read_file(path: &str) -> String {
     input
 }
 
-pub fn print_answer(day: u8, is_second: bool, answer: Option<u32>) {
+pub fn print_answer(day: u8, is_second: bool, answer: Option<usize>) {
     let part = if is_second { "Part2" } else { "Part1" };
     match answer {
         Some(answer) => println!("[AoC 2020 Day{day:02} {part}] {answer}"),
